requirements.txt
aws-cdk-lib==2.224.0
aws-cdk-aws-lambda-python-alpha==2.224.0a0
constructs>=10.0.0,<11.0.0
requests>=2.28.0
boto3>=1.26.0
